+++ src/parser/scanner.ts
@@ -38,7 +38,7 @@ export class Scanner implements ITokenStream {
 	/**
 	 * カーソル位置にあるトークンの種類を取得します。
 	*/
-	public get kind(): TokenKind {
+	public getKind(): TokenKind {
 		return this.token.kind;
 	}
 
@@ -74,8 +74,8 @@ export class Scanner implements ITokenStream {
 	 * 一致しなかった場合には文法エラーを発生させます。
 	*/
 	public expect(kind: TokenKind): void {
-		if (this.kind !== kind) {
-			throw new AiScriptSyntaxError(`unexpected token: ${TokenKind[this.kind]}`, this.token.loc);
+		if (this.getKind() !== kind) {
+			throw new AiScriptSyntaxError(`unexpected token: ${TokenKind[this.getKind()]}`, this.token.loc);
 		}
 	}